to try FileSystem APIs 1. # cordova plugin add cordova-plugin-file Fetching plugin "cordova-plugin-file" via npm Installing "cordova-plugin-file" for android
The Android Persistent storage location now defaults to "Internal".
Please check this plugins README to see if you application needs any changes in its config.xml.
If this is a new application no changes are required.
If this is an update to an existing application that did not specify an "AndroidPersistentFileLocation" you may need to add:
      "<preference name="AndroidPersistentFileLocation" value="Compatibility" />"
to config.xml in order for the application to find previously stored files.